+++ app/javascript/vue/tasks/otu/browse/components/timeline/Timeline.vue
@@ -32,7 +32,8 @@
             :key="item"
           >
             <li
-              v-show="filterSource(nomenclature.sources.list[item])">
+              v-show="filterSource(nomenclature.sources.list[item])"
+              class="horizontal-left-content">
               <label>
                 <input
                   v-model="references"
@@ -40,6 +41,7 @@
                   type="checkbox">
                 <span v-html="nomenclature.sources.list[item].cached"/>
               </label>
+              <radial-annotator :global-id="item"/>
             </li>
           </template>
         </template>
@@ -47,7 +49,9 @@
           <template
             v-for="(item, key) in nomenclature.sources.list"
             :key="key">
-            <li v-show="filterSource(item)">
+            <li
+              v-show="filterSource(item)"
+              class="horizontal-left-content">
               <label>
                 <input
                   v-model="references"
@@ -63,6 +67,7 @@
                   :style="{ 'background-color': topic.css_color }"
                   v-html="topic.name"/>
               </template>
+              <radial-annotator :global-id="key"/>
             </li>
           </template>
         </template>
@@ -196,6 +201,7 @@ import YearPicker from './TimelineYearsPick.vue'
 import extendSection from '../shared/extendSections'
 import SoftValidationModal from '../softValidationModal'
 import TimelineCitations from './TimelineCitations.vue'
+import RadialAnnotator from 'components/radials/annotator/annotator.vue'
 import { GetterNames } from '../../store/getters/getters'
 import { MutationNames } from '../../store/mutations/mutations'
 import { Otu } from 'routes/endpoints'
@@ -207,7 +213,8 @@ export default {
     ModalComponent,
     YearPicker,
     SoftValidationModal,
-    TimelineCitations
+    TimelineCitations,
+    RadialAnnotator
   },
   computed: {
     selectedReferences () {
@@ -344,9 +351,6 @@ export default {
   .topic-section {
     overflow-y: scroll;
     max-height: 480px;
-  }
-  li {
-    margin-top: 4px;
   }
   .references_topics {
     color: black;